Code adjustments suggested by ds5678 #2
