+++ BetterStacking/Patches.cs
@@ -1,22 +1,48 @@
 ﻿using HarmonyLib;
 using Il2Cpp;
-using MelonLoader;
+using System.Diagnostics.CodeAnalysis;
 using System.Reflection;
+using UnityEngine;
+
 namespace BetterStacking
 {
 
     internal static class Patches
     {
 
-        internal static bool PostFixTrack { get; set; } = false ;
-        internal static float PostfixCondition { get; set; }  = 0;
-        internal static GearItem? PostfixStack { get; set; }  = null;
-        internal static GearItem? PostfixGearToAdd { get; set; }  = null;
-        internal static float PostfixConstraint { get; set; }  = 0;
+        private static readonly string[] STACK_MERGE =
+        {
+        "GEAR_BirchSaplingDried",
+        "GEAR_BearHideDried",
+        "GEAR_BottleAntibiotics",
+        "GEAR_BottlePainKillers",
+        "GEAR_CoffeeTin",
+        "GEAR_GreenTeaPackage",
+        "GEAR_GutDried",
+        "GEAR_LeatherDried",
+        "GEAR_LeatherHideDried",
+        "GEAR_MapleSaplingDried",
+        "GEAR_MooseHideDried",
+        "GEAR_PackMatches",
+        "GEAR_RabbitPeltDried",
+        "GEAR_WolfPeltDried",
+        "GEAR_WoodMatches",
+        };
+
+        internal static bool PostFixTrack { get; private set; } = false;
+
+        internal static float PostfixCondition { get; private set; } = 0;
+
+        internal static GearItem? PostfixStack { get; private set; } = null;
+
+        internal static GearItem? PostfixGearToAdd { get; private set; } = null;
+
+        internal static float PostfixConstraint { get; private set; } = 0;
 
         [HarmonyPatch]
         internal static class PlayerManager_TryAddToExistingStackable
         {
+
             static MethodBase? TargetMethod()
             {
                 MethodInfo[] methods = typeof(PlayerManager).GetMethods();
@@ -27,12 +53,12 @@ namespace BetterStacking
                         return m;
                     }
                 }
-                MelonLogger.Warning("PlayerManager.TryAddToExistingStackable not found for patch.");
+                Implementation.LogWarning("PlayerManager.TryAddToExistingStackable not found for patch.");
                 return null;
             }
+
             internal static bool Prefix(ref GearItem gearToAdd, float normalizedCondition, int numUnits, out GearItem existingGearItem)
             {
-
                 existingGearItem = new GearItem();
 
                 if (normalizedCondition == 0)
@@ -40,7 +66,7 @@ namespace BetterStacking
                     return false;
                 }
 
-                if (Implementation.UseDefaultStacking(gearToAdd))
+                if (UseDefaultStacking(gearToAdd))
                 {
                     return true;
                 }
@@ -52,22 +78,22 @@ namespace BetterStacking
                     return false;
                 }
 
-
                 if (targetStack.m_StackableItem != null)
                 {
-                    Implementation.MergeIntoStack(normalizedCondition, numUnits, targetStack, gearToAdd);
+                    MergeIntoStack(normalizedCondition, numUnits, targetStack, gearToAdd);
                     existingGearItem = targetStack;
                     return true;
                 }
 
                 return false;
-
             }
+
         }
 
         [HarmonyPatch(typeof(ConsoleManager), nameof(ConsoleManager.CONSOLE_gear_add))]
         internal class ConsoleManager_CONSOLE_gear_add
         {
+
             private static void Postfix()
             {
                 // are we tracking a postfix patch ?
@@ -85,33 +111,13 @@ namespace BetterStacking
 
                 // reset the static values to avoid any conflicts
                 ResetPostfixParams();
-
             }
+
         }
 
-        internal static bool CanBeMerged(GearItem target, GearItem item)
+        internal static bool CanBeMerged([NotNullWhen(true)] GearItem? target, [NotNullWhen(true)] GearItem? item)
         {
-            return target != null && item != null && CanBeMerged(target.m_FlareItem, item.m_FlareItem);
-        }
-
-        private static bool CanBeMerged(FlareItem target, FlareItem item)
-        {
-            if (target == null || item == null)
-            {
-                return true;
-            }
-
-            if (target.IsBurning() || item.IsBurning())
-            {
-                return false;
-            }
-
-            if (target.IsBurnedOut() != item.IsBurnedOut())
-            {
-                return false;
-            }
-
-            return true;
+            return target != null && item != null;
         }
 
         internal static void ResetPostfixParams()
@@ -123,5 +129,94 @@ namespace BetterStacking
             PostfixConstraint = 0;
         }
 
+        internal static void MergeIntoStack(float normalizedCondition, int numUnits, GearItem targetStack, GearItem gearToAdd)
+        {
+            // check for console added items
+            if (uConsole.IsOn())
+            {
+                // normalizedCondition is always 1 here when added via console, this gets changed later in CONSOLE_gear_add
+                // so we recalculate it from the console params (or game logic) to be used in the below calculations
+
+                // NO condition specified (only 2 params)
+                if (uConsole.GetNumParameters() == 2)
+                {
+                    // calc default/random condition as per game logic
+                    gearToAdd.RollGearCondition(false);
+                    normalizedCondition = gearToAdd.GetNormalizedCondition();
+                }
+
+                // condition WAS specified (3rd param)
+                if (uConsole.GetNumParameters() == 3)
+                {
+                    // set the PostFixTrack to enable the CONSOLE_gear_add.postfix logic
+                    Patches.PostFixTrack = true;
+                    // calc condition based on console params
+                    float consoleCondition = Mathf.Clamp(float.Parse(uConsole.m_Argv[3]), 0, 100) / 100f;
+                    // apply the new condition and override normalizedCondition with the new value
+                    gearToAdd.CurrentHP = gearToAdd.m_GearItemData.m_MaxHP * consoleCondition;
+                    normalizedCondition = gearToAdd.GetNormalizedCondition();
+                }
+            }
+
+            int targetCount = numUnits + targetStack.m_StackableItem.m_Units;
+            float targetCondition = (numUnits * normalizedCondition + targetStack.m_StackableItem.m_Units * targetStack.GetNormalizedCondition()) / targetCount;
+
+            // set static variables for the postfox patch
+            if (Patches.PostFixTrack == true)
+            {
+                Patches.PostfixCondition = targetCondition;
+                Patches.PostfixStack = targetStack;
+                Patches.PostfixGearToAdd = gearToAdd;
+                Patches.PostfixConstraint = targetStack.m_StackableItem.m_StackConditionDifferenceConstraint;
+            }
+
+            // convince the game logic these stacks can merge between 0% -> 100% condition
+            // the game keeps the existing stack condition %
+            // (avoids having to destroy anything, we let the game do it all)
+            targetStack.m_StackableItem.m_StackConditionDifferenceConstraint = 100f;
+            gearToAdd.m_StackableItem.m_StackConditionDifferenceConstraint = 100f;
+
+            // change the target stack to the new calculated condition
+            targetStack.CurrentHP = targetCondition * targetStack.m_GearItemData.m_MaxHP;
+        }
+
+        internal static void SplitStack(GearItem gearItem)
+        {
+            if (gearItem == null || gearItem.m_StackableItem == null)
+            {
+                return;
+            }
+
+            try
+            {
+                SplitStackWithException(gearItem);
+            }
+            catch (Exception e)
+            {
+                Implementation.LogWarning("Failed to split stack of {gearItem.name}: {e}.");
+            }
+        }
+
+        internal static bool UseDefaultStacking(GearItem gearItem)
+        {
+            return gearItem == null || !STACK_MERGE.Contains(gearItem.name);
+        }
+
+        private static void SplitStackWithException(GearItem gearItem)
+        {
+            int count = gearItem.m_StackableItem.m_Units;
+            if (count <= 1)
+            {
+                return;
+            }
+
+            gearItem.m_StackableItem.m_Units = 1;
+
+            GearItem clone = GearItem.InstantiateGearItem(gearItem.name);
+            clone.m_StackableItem.m_Units = count - 1;
+            clone.CurrentHP = gearItem.CurrentHP;
+
+            GameManager.GetInventoryComponent().AddGear(clone);
+        }
     }
 }
